add item

+++ astra-db-java/src/test/java/com/datastax/astra/test/integration/local/LocalTableITTest.java
@@ -1,12 +1,17 @@
 package com.datastax.astra.test.integration.local;
 
 import com.datastax.astra.client.DataAPIClients;
+import com.datastax.astra.client.core.query.Filter;
+import com.datastax.astra.client.core.query.Projections;
+import com.datastax.astra.client.core.vector.DataAPIVector;
 import com.datastax.astra.client.databases.Database;
-import com.datastax.astra.client.tables.Table;
 import com.datastax.astra.client.tables.TableDefinition;
-import com.datastax.astra.client.tables.TableDescriptor;
 import com.datastax.astra.client.tables.TableOptions;
 import com.datastax.astra.client.tables.columns.ColumnTypes;
+import com.datastax.astra.client.tables.commands.TableDeleteResult;
+import com.datastax.astra.client.tables.commands.TableFindOneOptions;
+import com.datastax.astra.client.tables.commands.TableInsertManyOptions;
+import com.datastax.astra.client.tables.commands.TableInsertManyResult;
 import com.datastax.astra.client.tables.commands.TableInsertOneResult;
 import com.datastax.astra.client.tables.index.IndexDefinition;
 import com.datastax.astra.client.tables.index.IndexDefinitionOptions;
@@ -14,17 +19,34 @@ import com.datastax.astra.client.tables.index.IndexOptions;
 import com.datastax.astra.client.tables.index.VectorIndexDefinition;
 import com.datastax.astra.client.tables.index.VectorIndexDefinitionOptions;
 import com.datastax.astra.client.tables.index.VectorIndexOptions;
+import com.datastax.astra.client.tables.mapping.Column;
 import com.datastax.astra.client.tables.mapping.IntrospectedBean;
 import com.datastax.astra.client.tables.mapping.IntrospectedField;
+import com.datastax.astra.client.tables.mapping.Table;
 import com.datastax.astra.client.tables.row.Row;
 import com.datastax.astra.test.integration.AbstractTableITTest;
 import com.datastax.astra.test.model.TableSimpleRow;
 import com.dtsx.astra.sdk.db.domain.CloudProviderType;
 import com.dtsx.astra.sdk.utils.AstraEnvironment;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+import java.net.InetAddress;
+import java.net.UnknownHostException;
+import java.time.Duration;
+import java.time.Instant;
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 import static com.datastax.astra.client.core.query.Sorts.ascending;
@@ -71,9 +93,15 @@ public class LocalTableITTest extends AbstractTableITTest {
                     .dropIndex(INDEX_COUNTRY);
             getDatabase().dropTable(TABLE_SIMPLE);
         }
-        getDatabase().dropTable(TABLE_COMPOSITE);
-        getDatabase().dropTable(TABLE_ALL_RETURNS);
-        getDatabase().dropTable(TABLE_CASSIO);
+        if (getDatabase().tableExists(TABLE_CASSIO)) {
+            getDatabase().dropTable(TABLE_COMPOSITE);
+        }
+        if (getDatabase().tableExists(TABLE_ALL_RETURNS)) {
+            getDatabase().dropTable(TABLE_ALL_RETURNS);
+        }
+        if (getDatabase().tableExists(TABLE_CASSIO)) {
+            getDatabase().dropTable(TABLE_CASSIO);
+        }
         assertThat(getDatabase().tableExists(TABLE_SIMPLE)).isFalse();
         assertThat(getDatabase().tableExists(TABLE_COMPOSITE)).isFalse();
         assertThat(getDatabase().tableExists(TABLE_ALL_RETURNS)).isFalse();
@@ -84,7 +112,7 @@ public class LocalTableITTest extends AbstractTableITTest {
     @Order(2)
     public void shouldCreateTableSimple() {
         // Simple
-        Table<Row> tableSimple = getDatabase()
+        com.datastax.astra.client.tables.Table<Row> tableSimple = getDatabase()
                 .createTable(TABLE_SIMPLE, new TableDefinition()
                 .addColumnText("email")
                 .addColumnInt("age")
@@ -117,8 +145,8 @@ public class LocalTableITTest extends AbstractTableITTest {
 
     @Test
     @Order(4)
-    public void shouldCreateTableAllTypes() {
-        Table<Row> tableAllReturns = getDatabase().createTable(TABLE_ALL_RETURNS, new TableDefinition()
+    public void shouldCreateTableAllReturns() {
+        com.datastax.astra.client.tables.Table<Row> tableAllReturns = getDatabase().createTable(TABLE_ALL_RETURNS, new TableDefinition()
                 .addColumn("p_ascii", ColumnTypes.ASCII)
                 .addColumn("p_bigint", ColumnTypes.BIGINT)
                 .addColumn("p_blob", ColumnTypes.BLOB)
@@ -143,7 +171,7 @@ public class LocalTableITTest extends AbstractTableITTest {
                 .addColumnVector("p_vector", 3, COSINE)
                 .addColumnList("p_list_int", ColumnTypes.INT)
                 .addColumnSet("p_set_int", ColumnTypes.INT)
-                //.addColumnMap("p_map_text_text", ColumnTypes.TEXT, ColumnTypes.TEXT)
+                .addColumnMap("p_map_text_text", ColumnTypes.TEXT, ColumnTypes.TEXT)
                 .addColumn("p_double_minf", ColumnTypes.DOUBLE)
                 .addColumn("p_double_pinf", ColumnTypes.DOUBLE)
                 .addColumn("p_float_nan", ColumnTypes.FLOAT)
@@ -171,7 +199,6 @@ public class LocalTableITTest extends AbstractTableITTest {
     @Test
     @Order(5)
     public void shouldCreateTableAllCassio() {
-        getDatabase().dropTable(TABLE_CASSIO);
         assertThat(getDatabase().tableExists(TABLE_CASSIO)).isFalse();
         getDatabase().createTable(TABLE_CASSIO, new TableDefinition()
                         .addColumn("partition_id", ColumnTypes.TEXT)
@@ -209,20 +236,22 @@ public class LocalTableITTest extends AbstractTableITTest {
         String tableName = "table_simple";
         assertThat(getDatabase().tableExists(tableName)).isTrue();
         // Accessing the Table
-        Table<Row> tableSimple = getDatabase().getTable(tableName);
+        com.datastax.astra.client.tables.Table<Row> tableSimple = getDatabase().getTable(tableName);
         assertThat(tableSimple).isNotNull();
         assertThat(tableSimple.getTableName()).isEqualTo(tableName);
         // Accessing the Table Definitions
-        TableDescriptor desc = tableSimple.getDefinition();
+        TableDefinition desc = tableSimple.getDefinition();
         assertThat(desc).isNotNull();
-        assertThat(desc.getName()).isEqualTo(tableName);
-        assertThat(desc.getDefinition()).isNotNull();
-        assertThat(desc.getDefinition().getPrimaryKey().getPartitionBy()).isEqualTo(List.of("email"));
+        assertThat(desc.getPrimaryKey().getPartitionBy()).isEqualTo(List.of("email"));
     }
 
+    // ------------------------------------------
+    // InsertOne
+    // ------------------------------------------
+
     @Test
-    public void shouldMapTableSimple() {
-        Table<Row> table = getDatabase().getTable(TABLE_SIMPLE);
+    public void shouldInsertOneTableSimple() {
+        com.datastax.astra.client.tables.Table<Row> table = getDatabase().getTable(TABLE_SIMPLE);
         Row row = new Row()
                 .addBoolean("human", true)
                 .addInt("age", 42)
@@ -230,42 +259,200 @@ public class LocalTableITTest extends AbstractTableITTest {
                 .addText("country", "France")
                 .addText("email",  "cedrick@datastax.com");
         TableInsertOneResult res = table.insertOne(row);
-        System.out.println(res.insertedIds());
-
-        //Table<TableSimpleRow> table2 = getDatabase().getTable(TableSimpleRow.class);
-        //table2.insertOne(new TableSimpleRow());
-
+        System.out.println(res.getInsertedId());
+        System.out.println(res.getInsertedIdAsRow().get("email"));
     }
 
     @Test
-    public void shouldIntrospectBean() {
-        IntrospectedBean<TableSimpleRow> rowDecorator = new IntrospectedBean<>(TableSimpleRow.class);
-        System.out.println("Table Name: " + rowDecorator.getName());
-        for (IntrospectedField field : rowDecorator.getFields().values()) {
-            System.out.println("Field: " + field.getName() + ", Type: " + field.getType().getName());
-        }
-    }
-
-    @Test
-    public void shouldInsertOneAllReturns() {
-        Table<Row> tableAllReturns = getDatabase().getTable(TABLE_ALL_RETURNS);
+    public void shouldInsertOneTableComposite() {
+        com.datastax.astra.client.tables.Table<Row> table = getDatabase().getTable(TABLE_COMPOSITE);
         Row row = new Row()
-                .addText("p_ascii", "abc")
+                .addInt("age", 42)
+                .addText("name", "John")
+                .addText("id", "John");
+        TableInsertOneResult res = table.insertOne(row);
+        assertThat(res.getInsertedId().size()).isEqualTo(1);
+        assertThat(res.getInsertedIdAsRow().size()).isEqualTo(1);
+    }
+
+    @Data @NoArgsConstructor @AllArgsConstructor
+    private static class TableCompositeRow {
+        private int age;
+        private String name;
+        private String id;
+    }
+
+    @Test
+    public void shouldInsertOneTableCompositeBean() {
+        com.datastax.astra.client.tables.Table<TableCompositeRow> table = getDatabase().getTable(TABLE_COMPOSITE, TableCompositeRow.class);
+        TableCompositeRow row = new TableCompositeRow(42, "Cedrick", "Lunven");
+        TableInsertOneResult res = table.insertOne(row);
+        assertThat(res.getInsertedId().size()).isEqualTo(2);
+        assertThat(res.getInsertedIdAsRow().getText("name")).isEqualTo("Cedrick");
+    }
+
+    @Data
+    @Table
+    @NoArgsConstructor @AllArgsConstructor
+    private static class TableCompositeRowAnnotation {
+
+        @Column(name = "age", type = ColumnTypes.INT)
+        private int agex;
+
+        @Column(name = "name", type = ColumnTypes.TEXT)
+        private String namex;
+
+        @Column(name = "idx", type = ColumnTypes.TEXT)
+        private String idx;
+    }
+
+    @Test
+    public void shouldInsertOneTableAnnotatedBean() {
+        com.datastax.astra.client.tables.Table<TableCompositeRowAnnotation> table = getDatabase().getTable(TABLE_COMPOSITE, TableCompositeRowAnnotation.class);
+        TableCompositeRowAnnotation row = new TableCompositeRowAnnotation(42, "Cedrick", "Lunven");
+        TableInsertOneResult res = table.insertOne(row);
+        assertThat(res.getInsertedId().size()).isEqualTo(1);
+        assertThat(res.getInsertedIdAsRow().size()).isEqualTo(1);
+    }
+
+    // ------------------------------------------
+    // InsertMany
+    // ------------------------------------------
+
+    @Test
+    public void shouldInsertManyTableComposite() {
+        com.datastax.astra.client.tables.Table<TableCompositeRow> table = getDatabase().getTable(TABLE_COMPOSITE, TableCompositeRow.class);
+        TableCompositeRow row = new TableCompositeRow(42, "Cedrick", "Lunven");
+        TableInsertManyResult res = table.insertMany(List.of(
+                new TableCompositeRow(42, "Cedrick", "Lunven"),
+                new TableCompositeRow(41, "Hind", "Lunven")),
+                new TableInsertManyOptions());
+    }
+
+
+    @Test
+    public void shouldFindOneTableComposite() {
+        com.datastax.astra.client.tables.Table<Row> table = getDatabase().getTable(TABLE_COMPOSITE);
+        Row row = new Row()
+                .addInt("age", 42)
+                .addText("name", "John")
+                .addText("id", "John");
+        Filter johnFilter = new Filter(Map.of("id", "John","name", "John"));
+        Optional<Row> res = table.findOne(johnFilter,
+                new TableFindOneOptions().projection(Projections.include("id", "age")));
+        assertThat(res).isPresent();
+    }
+
+    @Test
+    public void shouldDeleteOneTableComposite() {
+        com.datastax.astra.client.tables.Table<Row> table = getDatabase().getTable(TABLE_COMPOSITE);
+        Row row = new Row()
+                .addInt("age", 42)
+                .addText("name", "John")
+                .addText("id", "John");
+        Filter johnFilter = new Filter(Map.of("id", "John","name", "John"));
+        TableDeleteResult res = table.deleteOne(johnFilter);
+        assertThat(res.getDeletedCount()).isEqualTo(1);
+
+    }
+
+
+
+    @Test
+    public void shouldInsertOneAllReturns() throws UnknownHostException {
+        String timeString = "13:30:54.234";
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("HH:mm:ss.SSS");
+        LocalTime localTime = LocalTime.parse(timeString, formatter);
+
+        Row row = new Row()
+                .addAscii("p_ascii", "abc")
                 .addBigInt("p_bigint", 10002L)
                 .addInt("p_int", 987)
                 .addBoolean("p_boolean", false)
                 .addText("p_text", "Ålesund")
-                .addDouble("p_double_pinf", Double.MAX_VALUE);
-                //.addBlob("p_blob", "blob".getBytes());
-                //.addDate("p_date", LocalDate.now())
-                //.addDecimal("p_decimal", new BigDecimal("42.42"))
-                //.addByte("p_tinyint", (byte) 42)
-                //.addDouble("p_double", 987.6543)
-                //.addDuration("p_duration", Duration.ofSeconds(42))
-                //.addFloat("p_float", 42.42f)
-                //.addInt("p_int", 42)
-                //.addInet("p_inet", InetAddress.getByName("
+                .addText("p_text_omitted", null)
+                .addDouble("p_double_pinf", Double.MAX_VALUE)
+                .addDouble("p_double_minf", Double.NEGATIVE_INFINITY)
+                .addBlob("p_blob", "blob".getBytes())
+                .addSmallInt("p_smallint", (short) 200)
+                .addVarInt("p_varint",new BigInteger("444"))
+                .addTinyInt("p_tinyint",(byte) 17)
+                .addDuration("p_duration", Duration.ofHours(12).plusMinutes(48))
+                .addInet("p_inet", InetAddress.getByAddress(new byte[]{12, 34, 56, 78}))
+                .addDouble("p_double", 987.6543d)
+                .addFloat("p_float", 66.55f)
+                .addFloat("p_float_nan", Float.NaN)
+                .addTimeStamp("p_timestamp", Instant.now())
+                .addTime("p_time", localTime)
+                .addUUID("p_uuid", java.util.UUID.fromString("9c5b94b1-35ad-49bb-b118-8e8fc24abf80"))
+                .addDate("p_date", LocalDate.of(2015,5,3))
+                .addDecimal("p_decimal", new BigDecimal("123.45"))
+                .addVector("p_vector", DataAPIVector.of(.1f, 0.2f, 0.3f))
+                .addList("p_list_int", List.of(4, 17, 34))
+                .addSet("p_set_int",  Set.of(9, 81));
+
+                // .add("p_vector", new float[] {.1f, .2f, .3f})
+                //.addTableDuration("p_duration", TableDuration.of(
+                //        Period.ofDays(3),
+                //        Duration.ofHours(12).plusMinutes(48)));
         getDatabase().getTable(TABLE_ALL_RETURNS).insertOne(row);
     }
+
+
+    @Test
+    public void shouldFindOneAllReturns() throws UnknownHostException {
+        com.datastax.astra.client.tables.Table<Row> tableAllReturns = getDatabase().getTable(TABLE_ALL_RETURNS);
+
+        Filter top = new Filter(Map.of(
+                "p_ascii", "abc",
+                "p_bigint", 10002L));
+        Row row = tableAllReturns
+                .findOne(top)
+                .orElseThrow(() -> new IllegalArgumentException("Row not found"));
+        assertThat(row.getText("p_ascii")).isEqualTo( "abc");
+        assertThat(row.getBigInt("p_bigint")).isEqualTo( 10002L);
+
+        //Row row2 = tableAllReturns.findOne(
+        //        and(eq( "p_ascii", "abc"),
+        //            eq("p_bigint", 10002L)
+        //        )).orElseThrow(() -> new IllegalArgumentException("Row not found"));
+
+        //assertThat(row2.getText("p_ascii")).isEqualTo( "abc");
+        //assertThat(row2.getBigInt("p_bigint")).isEqualTo( 10002L);
+        /*
+        .addAscii("p_ascii", "abc")
+                .addBigInt("p_bigint", 10002L)
+                .addInt("p_int", 987)
+                .addBoolean("p_boolean", false)
+                .addText("p_text", "Ålesund")
+                .addText("p_text_omitted", null)
+                .addDouble("p_double_pinf", Double.MAX_VALUE)
+                .addDouble("p_double_minf", Double.NEGATIVE_INFINITY)
+                .addBlob("p_blob", "blob".getBytes())
+                .addSmallInt("p_smallint", (short) 200)
+                .addVarInt("p_varint",444)
+                .addTinyInt("p_tinyint",(short) 17)
+                .addDuration("p_duration", Duration.ofHours(12).plusMinutes(48))
+                .addInet("p_inet", InetAddress.getByAddress(new byte[]{12, 34, 56, 78}))
+                .addDouble("p_double", 987.6543d)
+                .addFloat("p_float", 66.55f)
+                .addFloat("p_float_nan", Float.NaN)
+                .addTimeStamp("p_timestamp", Instant.now())
+                .addTime("p_time", localTime)
+                .addUUID("p_uuid", java.util.UUID.fromString("9c5b94b1-35ad-49bb-b118-8e8fc24abf80"))
+                .addDate("p_date", LocalDate.of(2015,5,3))
+                .addDecimal("p_decimal", new BigDecimal("123.45"))
+                .addVector("p_vector", DataAPIVector.of(.1f, 0.2f, 0.3f))
+                .addList("p_list_int", List.of(4, 17, 34))
+                .addSet("p_set_int",  Set.of(9, 81));
+
+                //.addTableDuration("p_duration", TableDuration.of(
+                //        Period.ofDays(3),
+                //        Duration.ofHours(12).plusMinutes(48)));
+         */
+    }
+
+
+
 
 }
